requirements.txt
Django == 4.0.2
django-cors-headers
djangorestframework
pillow
psycopg2
selenium
djangorestframework-simplejwt
django-extensions
<<<<<<< HEAD
celery
redis
django-celery-results
django-celery-beat
=======
BeautifulSoup4
>>>>>>> master
